🛠️ Build Dependabot updates (#298)
All of our TypeScript builds have failed in this project because we
check to see if the `dist/` directory matches what is in the pull
request. Dependabot will open a pull request with changes to the
manifest/lock files, but it won't build the dependencies. By building
them and committing, we should be able to pass the check we're
enforcing.

Note: I'm using `[dependabot skip]` in the commit message as it will
allow Dependabot to rebase/recreate changes. Without that, the PR will
be left open (and new changes will not be proposed).

---------

Co-authored-by: Samuel Giddins <segiddins@segiddins.me>