Ensure we're not in a detached state when committing (#304)
When running this workflow for the first time, we got the following
output:
```
fatal: You are not currently on a branch.
To push the history leading to the current (detached HEAD)
state now, use

    git push origin HEAD:<name-of-remote-branch>
```

By specifying the ref on checkout, we should be in a valid state and be
able to push.